refactor(DAO-2005): migrate my-rewards date formatters to shared utilities
Remove formatExpandedDate, formatDateRange, and formatDateForCsv from
tx-history/utils. Consumers now import from @/lib/utils; formatDateRange
callers pass cycleDuration.as('seconds') instead of a Duration object.
